Adding Javadoc and separating TestRationalNumberEvaluator from TestEvaluator
