+++ src/main/java/calculator/Divides.java
@@ -49,6 +49,12 @@ public final class Divides extends Operation
   public int op(int l, int r)
     { return (l/r); }
 
+    /**
+     * The actual computation of the (binary) arithmetic division of two rational numbers
+     * @param l The first rational number
+     * @param r The second rational number that should divide the first
+     * @return The rational number that is the result of the division
+     */
     @Override
     public MyRationalNumber op(MyRationalNumber l, MyRationalNumber r) {
         return l.divide(r);
