+++ src/main/java/calculator/Minus.java
@@ -50,6 +50,12 @@ public final class Minus extends Operation
   	return (l-r);
   }
 
+     /**
+      * The actual computation of the (binary) arithmetic subtraction of two rational numbers
+      * @param l	 The first rational number
+      * @param r     The second rational number that should be subtracted from the first
+      * @return
+      */
      @Override
      public MyRationalNumber op(MyRationalNumber l, MyRationalNumber r) {
          return l.subtract(r);
