+++ src/main/java/calculator/MyNumber.java
@@ -19,7 +19,7 @@ public abstract class MyNumber implements Expression
     public abstract Integer getInteger();
 
     /**
-     * abstract method to obtain the rational number contain in the object
+     * abstract method to obtain the rational number corresponding to the value of the object
      * @return The rational number contain in the object
      */
     public abstract MyRationalNumber getRational();
