+++ src/main/java/calculator/Times.java
@@ -48,6 +48,13 @@ public final class Times extends Operation
   public int op(int l, int r)
     { return (l*r); }
 
+     /**
+      * The actual computation of the (binary) arithmetic multiplication of two rational numbers
+      * @param l The first rational number
+      * @param r The second rational number that should be multiplied with the first
+      * @return The rational number that is the result of the multiplication
+      */
+
      @Override
      public MyRationalNumber op(MyRationalNumber l, MyRationalNumber r) {
          return l.multiply(r);
