+++ src/test/java/calculator/TestEvaluator.java
@@ -13,17 +13,12 @@ class TestEvaluator {
 
     private Calculator calc;
     private int value1, value2;
-    private int nominator1, nominator2, denominator1, denominator2;
 
     @BeforeEach
     void setUp() {
         calc = new Calculator();
         value1 = 8;
         value2 = 6;
-        nominator1 = 1;
-        nominator2 = 2;
-        denominator1 = 3;
-        denominator2 = 4;
 
     }
 
@@ -32,10 +27,6 @@ class TestEvaluator {
         assertEquals( value1, calc.eval(new MyInteger(value1)));
     }
 
-    @Test
-    void testEvaluatorRationalNumber() {
-        assertEquals( MyRationalNumber.create(nominator1, denominator1), calc.evalRational(MyRationalNumber.create(nominator1, denominator1)));
-    }
 
     @ParameterizedTest
     @ValueSource(strings = {"*", "+", "/", "-"})
@@ -56,42 +47,5 @@ class TestEvaluator {
         }
     }
 
-    @ParameterizedTest
-    @ValueSource(strings = {"*", "+", "/", "-"})
-    void testEvaluateRationalOperations(String symbol) {
-        List<Expression> params = Arrays.asList(MyRationalNumber.create(nominator1, denominator1),MyRationalNumber.create(nominator2, denominator2));
-        try {
-            //construct another type of operation depending on the input value
-            //of the parameterised test
-            switch (symbol) {
-                case "+"	->	assertEquals( MyRationalNumber.create(5,6), calc.evalRational(new Plus(params)));
-                case "-"	->	assertEquals( MyRationalNumber.create(-1, 6), calc.evalRational(new Minus(params)));
-                case "*"	->	assertEquals( MyRationalNumber.create(1, 6), calc.evalRational(new Times(params)));
-                case "/"	->	assertEquals( MyRationalNumber.create(2, 3), calc.evalRational(new Divides(params)));
-                default		->	fail();
-            }
-        } catch (IllegalConstruction e) {
-            fail();
-        }
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {"*", "+", "/", "-"})
-    void testEvaluateMixedOperations(String symbol) {
-        List<Expression> params = Arrays.asList(MyRationalNumber.create(nominator1, denominator1),MyRationalNumber.create(nominator2, denominator2), new MyInteger(value1));
-        try {
-            //construct another type of operation depending on the input value
-            //of the parameterised test
-            switch (symbol) {
-                case "+"	->	assertEquals( MyRationalNumber.create(53,6), calc.evalRational(new Plus(params)));
-                case "-"	->	assertEquals( MyRationalNumber.create(-49, 6), calc.evalRational(new Minus(params)));
-                case "*"	->	assertEquals( MyRationalNumber.create(8, 6), calc.evalRational(new Times(params)));
-                case "/"	->	assertEquals( MyRationalNumber.create(2, 24), calc.evalRational(new Divides(params)));
-                default		->	fail();
-            }
-        } catch (IllegalConstruction e) {
-            fail();
-        }
-    }
 
 }
